refactor: move PG state-proxy from asya-gateway to asya-state-proxy/go
State proxy implementations belong in src/asya-state-proxy/, organized
by backend type. The PG connector was incorrectly placed inside
asya-gateway.

- New Go module: asya-state-proxy-go (src/asya-state-proxy/go/)
- Package renamed: stateproxypg → pg (internal/pg/)
- Separate Dockerfile.pg for the PG connector binary
- Component test compose updated to build from new location
- Gateway Dockerfile/Makefile cleaned of state-proxy-pg references
- No cross-module Go imports (HTTP over Unix socket only)
