+++ src/asya-state-proxy/Dockerfile.pg
@@ -1,0 +1,21 @@
+FROM golang:1.25-alpine AS builder
+WORKDIR /app/go
+
+# Copy go mod files and download dependencies
+COPY go/go.mod go/go.sum ./
+RUN --mount=type=cache,target=/go/pkg/mod \
+    go mod download
+
+# Copy source
+COPY go/ .
+
+# Build binary
+RUN --mount=type=cache,target=/root/.cache/go-build \
+    --mount=type=cache,target=/go/pkg/mod \
+    CGO_ENABLED=0 GOOS=linux go build -o /state-proxy-pg ./cmd/state-proxy-pg/
+
+# Runtime image
+FROM alpine:3.20
+RUN apk --no-cache add ca-certificates
+COPY --from=builder /state-proxy-pg /state-proxy-pg
+ENTRYPOINT ["/state-proxy-pg"]
