+++ src/asya-state-proxy/Makefile
@@ -2,7 +2,9 @@
 
 test-unit:
 	uv run --project . --extra test --extra s3 --extra gcs --extra redis pytest tests/ -v
+	$(MAKE) -C go test-unit
 
 lint:
 	uv run --project . ruff check . --fix
 	uv run --project . ruff format .
+	$(MAKE) -C go lint
