+++ src/asya-state-proxy/go/internal/pg/connector_test.go
@@ -1,4 +1,4 @@
-package stateproxypg
+package pg
 
 import (
 	"context"
