+++ src/asya-state-proxy/go/internal/pg/query.go
@@ -1,4 +1,4 @@
-package stateproxypg
+package pg
 
 import (
 	"context"
