+++ src/asya-state-proxy/go/internal/pg/query_test.go
@@ -1,4 +1,4 @@
-package stateproxypg
+package pg
 
 import (
 	"testing"
