+++ src/asya-state-proxy/go/internal/pg/schema.go
@@ -1,4 +1,4 @@
-package stateproxypg
+package pg
 
 import (
 	"context"
