+++ src/asya-state-proxy/go/internal/pg/server_test.go
@@ -1,4 +1,4 @@
-package stateproxypg
+package pg
 
 import (
 	"encoding/json"
